fixed ServerSideLeafGuessTest.

+++ components/gnutella-core/src/test/java/com/limegroup/gnutella/ServerSideLeafGuessTest.java
@@ -358,7 +358,7 @@ public class ServerSideLeafGuessTest extends ClientSideTestCase {
     }
 
     public static Integer numUPs() {
-        return new Integer(4);
+        return new Integer(3);
     }
 
     public static ActivityCallback getActivityCallback() {
